Move locationapi to standalone class
